requirements.txt
fastapi~=0.115.7
uvicorn[standard]~=0.34.0
SQLAlchemy~=2.0.37
pydantic-settings~=2.7.1
pydantic~=2.10.6
h11~=0.14.0
pip~=24.3.1
Mako~=1.3.8
MarkupSafe~=3.0.2
typing_extensions~=4.12.2
exceptiongroup~=1.2.2
sniffio~=1.3.1
click~=8.1.8
alembic~=1.14.1
asyncpg~=0.30.0
async-timeout~=5.0.1
starlette~=0.45.3
pydantic_core~=2.27.2
anyio~=4.8.0
greenlet~=3.1.1
annotated-types~=0.7.0
setuptools~=59.6.0
python-dotenv~=1.0.1
cryptography~=44.0.0
rsa~=4.9
pyasn1~=0.6.1
cffi~=1.17.1
python-jose~=3.3.0
ecdsa~=0.19.0
six~=1.17.0
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
certifi~=2024.12.14
pycparser~=2.22
email-validator~=2.2.0

dnspython~=2.7.0
pytest~=8.3.4
pluggy~=1.5.0
iniconfig~=2.0.0
email_validator~=2.2.0
packaging~=24.2
tomli~=2.2.1
passlib[bcrypt]
pytest-asyncio~=0.21.0
pytest-cov~=4.1.0
pytest-mock~=3.10.0
APScheduler==3.10.1